moving the galewsky test case inital conditions to a separate file for use in both the explicit and rosenbrock solvers
